Implementation of unsub-ing listener from data provider.
